Adding possibility to configure sass generation directories
By default "sass" files should be stored in a "sass" directory in stylesheet's directory and generated "css" files are saved in stylesheet's directory.
The idea is not to have generated "css" stored in same directory as the others and to allow "sass" to be stored in the root stylesheet's directory of themes.
Both directory can be set in config as "css_cache_dir" and "sass_dir"
eg: all "css" can be saved in a specific cache directory "config.css_cache_dir = :root/tmp/cache/themes_for_rails/:name/stylesheets"